Create biggest_digits
Greatest numbers 🌶️🌶️
The input to the program is a natural number nn, and then nn different natural numbers, each on a separate line. Write a program that prints the largest and second largest numbers in a sequence.

Input data format
The input to the program is a natural number n \ ge 2n≥2, and then nn different natural numbers, each on a separate line.

Output data format
The program should print the two largest numbers, each on a separate line.